Move git related logic into separated file
